Changed the structure of the sentence.
I'm not sure if it's only me (English is not my first language) but this sentence messed up my mind. I understood from the sentence that this requirement may disappear one day. To make it easier for other non-English speakers I changed the sentence structure.

Before: And it is not impossible that this requirement will one day disappear.

After: And this requirement may one day disappear.

Or we could do something like this: And it is possible that this requirement will one day disappear.