Grammatical and stylistic modifications to docs (#3103)
I made grammatical and stylistic modifications to the `Essentials`
portion of the documentation.

--

I was reading through the documentation for the first time as a new user
and thought some of my changes could be useful. If I was overzealous, I
apologize.

I'm excited to begin using the `riverpod` package in my project. Thank
you the time that's been spent on this project.